+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * 【core   】      `PathUtil#del`增加null检查（pr#1331@Gitee）
 * 【db     】      增加SAP HANA识别及方言（pr#3914@Github）
 * 【crypto 】      增加`Argon2`类，实现Argon2算法（issue#3890@Github）
+* 【core   】      `CharSequenceUtil`增加toLoweCase和toUpperCase方法（issue#IC0H2B@Gitee）
 
 ### 🐞Bug修复